refactor overload filter execute method
