+++ src/main/java/use_case/game_filter/GameFilterInputBoundary.java
@@ -6,11 +6,6 @@ package use_case.game_filter;
 public interface GameFilterInputBoundary {
 
     /**
-     * Execute the game filter Use Case.
-     */
-    void execute();
-
-    /**
      * Execute the game filter Use Case and process the input data.
      * @param gameFilterInputData the input data for this use case
      */
